refactor: changed location of some dtos in auth and removed AuthResponse
