--- auth/src/main/java/com/github/zigcat/merchsite_microservice/auth/dto/AuthResponse.java
@@ -1,14 +1,0 @@
-package com.github.zigcat.merchsite_microservice.auth.dto;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthResponse {
-    private String status;
-    private JwtResponse response;
-}
